pyproject.toml
[tool.poetry]
name = "gpt-server"
version = "0.0.1"
description = ""
authors = ["liuyu <506610466@qq.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.350"
sentence-transformers = "^2.2.2"
fschat = {extras = ["model-worker", "webui"], version = "^0.2.34"}
loguru = "^0.7.2"
deepspeed = "^0.12.6"
mpi4py = "^3.1.5"
distribute = "^0.7.3"
wheel = "^0.42.0"
transformers-stream-generator = "^0.0.4"

[[tool.poetry.source]]
name = "ustc"
url = "https://mirrors.ustc.edu.cn/pypi/web/simple"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
